generator/agents/tests/test_docgen_response_handler.py
"""
test_docgen_response_handler.py
Enterprise-grade test suite for the DocGen Response Handler module.

This test suite provides comprehensive coverage including:
- Unit tests for all components
- Integration tests for the full pipeline
- Performance and load testing
- Security and compliance testing
- Error handling and edge cases
- Mock services and fixtures
"""

import asyncio
import json
import os
import tempfile
import time
import uuid
from datetime import datetime
from pathlib import Path
from typing import Any, Dict, List, Optional
from unittest.mock import AsyncMock, MagicMock, Mock, patch, call
import concurrent.futures
import random
import string

import pytest
import pytest_asyncio
from pytest_benchmark.fixture import BenchmarkFixture
import aiohttp
from aiohttp import web
import aiofiles
from faker import Faker
from hypothesis import given, strategies as st, settings, assume
from hypothesis.stateful import RuleBasedStateMachine, rule, initialize, Bundle
import pypandoc
import docutils.core
from prometheus_client import REGISTRY
from opentelemetry import trace
from opentelemetry.sdk.trace import TracerProvider
from opentelemetry.sdk.trace.export import SimpleSpanProcessor
from opentelemetry.sdk.trace.export.in_memory_span_exporter import InMemorySpanExporter

# Import the module under test
from agents.docgen_response_handler import (
from docgen_response_handler import (
    ResponseHandler,
    FormatPlugin,
    MarkdownPlugin,
    reStructuredTextPlugin,
    HTMLPlugin,
    PluginRegistry,
    scrub_text,
    scan_content_for_findings,
    REQUIRED_SECTIONS,
    DANGEROUS_CONTENT_PATTERNS,
    process_calls_total,
    process_errors_total,
    process_latency_seconds,
    content_quality_score,
    security_findings_gauge,
    section_status_gauge
)

# Initialize faker for test data generation
fake = Faker()

# Configure OpenTelemetry for testing
memory_exporter = InMemorySpanExporter()
span_processor = SimpleSpanProcessor(memory_exporter)
tracer_provider = TracerProvider()
tracer_provider.add_span_processor(span_processor)
trace.set_tracer_provider(tracer_provider)


# ============================================================================
# FIXTURES AND MOCKS
# ============================================================================

@pytest.fixture
def mock_presidio():
    """Mock Presidio analyzer and anonymizer for testing."""
    with patch('docgen_response_handler.AnalyzerEngine') as mock_analyzer_class, \
         patch('docgen_response_handler.AnonymizerEngine') as mock_anonymizer_class:
        
        mock_analyzer = MagicMock()
        mock_anonymizer = MagicMock()
        
        mock_analyzer_class.return_value = mock_analyzer
        mock_anonymizer_class.return_value = mock_anonymizer
        
        # Configure analyze to return sample results
        mock_analyzer.analyze.return_value = [
            MagicMock(entity_type="EMAIL_ADDRESS", start=10, end=25),
            MagicMock(entity_type="PHONE_NUMBER", start=30, end=42)
        ]
        
        # Configure anonymize to return redacted text
        mock_anonymizer.anonymize.return_value = MagicMock(
            text="Sample text with [REDACTED] and [REDACTED] information."
        )
        
        yield mock_analyzer, mock_anonymizer


@pytest.fixture
def mock_llm():
    """Mock LLM for testing repair and correction features."""
    with patch('docgen_response_handler.generate_docs_llm') as mock_gen:
        async def mock_llm_response(*args, **kwargs):
            prompt = args[0] if args else kwargs.get('prompt', '')
            
            # Return contextual responses based on prompt content
            if "missing" in prompt.lower():
                return {
                    "content": "# Introduction\nFixed introduction.\n## Installation\n`pip install`\n## Usage\nUsage guide.\n## License\nMIT License.",
                    "output_tokens": 50,
                    "usage": {"prompt_tokens": 100, "completion_tokens": 50}
                }
            elif "malformed" in prompt.lower():
                return {
                    "content": "# Corrected Document\n\nThis is properly formatted content.",
                    "output_tokens": 20,
                    "usage": {"prompt_tokens": 80, "completion_tokens": 20}
                }
            else:
                return {
                    "content": "Generic LLM response.",
                    "output_tokens": 10,
                    "usage": {"prompt_tokens": 50, "completion_tokens": 10}
                }
        
        mock_gen.side_effect = mock_llm_response
        yield mock_gen


@pytest.fixture
def mock_plantuml():
    """Mock PlantUML for diagram generation testing."""
    with patch('docgen_response_handler.PlantUML') as mock_plantuml_class:
        mock_plantuml = MagicMock()
        mock_plantuml_class.return_value = mock_plantuml
        mock_plantuml.get_url.return_value = "https://plantuml.com/mock-diagram.png"
        yield mock_plantuml


@pytest.fixture
async def temp_repo():
    """Create a temporary Git repository for testing."""
    with tempfile.TemporaryDirectory() as temp_dir:
        repo_path = Path(temp_dir)
        
        # Initialize Git repo
        proc = await asyncio.create_subprocess_exec(
            "git", "init",
            cwd=repo_path,
            stdout=asyncio.subprocess.PIPE,
            stderr=asyncio.subprocess.PIPE
        )
        await proc.communicate()
        
        # Configure Git
        for cmd in [
            ["git", "config", "user.email", "test@example.com"],
            ["git", "config", "user.name", "Test User"]
        ]:
            proc = await asyncio.create_subprocess_exec(
                *cmd,
                cwd=repo_path,
                stdout=asyncio.subprocess.PIPE,
                stderr=asyncio.subprocess.PIPE
            )
            await proc.communicate()
        
        # Create test files
        test_file = repo_path / "test.py"
        async with aiofiles.open(test_file, 'w') as f:
            await f.write("def test_function():\n    pass\n")
        
        # Commit
        for cmd in [
            ["git", "add", "."],
            ["git", "commit", "-m", "Initial commit"]
        ]:
            proc = await asyncio.create_subprocess_exec(
                *cmd,
                cwd=repo_path,
                stdout=asyncio.subprocess.PIPE,
                stderr=asyncio.subprocess.PIPE
            )
            await proc.communicate()
        
        yield str(repo_path)


@pytest.fixture
def sample_markdown_content():
    """Generate sample Markdown content for testing."""
    return """# Sample Documentation

## Introduction
This is a sample documentation for testing purposes.

## Installation
```bash
pip install sample-package