generator/master_requirements.txt
aioboto3==15.1.0
aiobotocore==2.24.0
aiobreaker==1.2.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-client-cache==0.13.0
aioitertools==0.12.0
aiokafka==0.12.0
aiolimiter==1.2.1
aioresponses==0.7.8
aiormq==6.9.0
aiosignal==1.4.0
aiosmtplib==4.0.2
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.64.0
anyio==4.10.0
anytree==2.13.0
APScheduler==3.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.9.1
astroid==3.3.11
async-timeout==4.0.3
asyncpg==0.30.0
attrs==25.3.0
azure-core==1.35.0
azure-identity==1.24.0
azure-monitor-ingestion==1.1.0
azure-monitor-query==2.0.0
backoff==2.2.1
backports-datetime-fromisoformat==2.0.3
backports.asyncio.runner==1.2.0
bandit==1.8.6
bcrypt==4.3.0
bigtree==0.30.1
bitarray==3.7.1
black==25.1.0
bleach==6.2.0
blinker==1.9.0
blis==1.3.0
boto3==1.39.11
botocore==1.39.11
cachetools==5.5.2
catalogue==2.0.10
Cerberus==1.3.7
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
circuitbreaker==2.1.3
ckzg==2.1.1
click==8.2.1
cloudpathlib==0.22.0
cloudpickle==3.1.1
colorama==0.4.6
confection==0.1.5
contourpy==1.3.2
coverage==7.10.6
cryptography==44.0.3
cycler==0.12.1
cymem==2.0.11
cytoolz==1.0.1
dask==2025.9.0
dataclasses-json==0.6.7
deap==1.4.3
Deprecated==1.2.18
dill==0.3.9
distro==1.9.0
docker==7.1.0
durationpy==0.10
ecdsa==0.19.1
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl#sha256=293e9547a655b25499198ab15a525b05b9407a75f10255e405e8c3854329ab63
environs==14.3.0
etcd3-py==0.1.6
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
exceptiongroup==1.3.0
Faker==37.6.0
Farama-Notifications==0.0.4
fastapi==0.116.1
fastapi-cache2==0.2.2
feast==0.53.0
ffmpeg==1.4
filelock==3.19.1
flake8==7.3.0
Flask==3.1.2
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
flask-swagger-ui==5.21.0
fonttools==4.59.2
freezegun==1.5.5
frozenlist==1.7.0
fsspec==2025.7.0
future-fstrings==1.2.0
gitdb==4.0.12
GitPython==3.1.45
Gnosis==0.1.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-pubsub==2.25.0
google-cloud-secret-manager==2.24.0
google-cloud-storage==3.3.0
google-crc32c==1.7.1
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.21
greenlet==3.2.4
grpc-google-iam-v1==0.14.2
grpcio==1.74.0
grpcio-health-checking==1.63.2
grpcio-status==1.63.2
grpcio-tools==1.63.2
gunicorn==23.0.0
gymnasium==1.2.0
h11==0.16.0
hexbytes==1.3.1
hiredis==3.2.1
httpcore==1.0.9
httplib2==0.31.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
hvac==2.3.0
hypothesis==6.144.0
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.5.2
influxdb-client==1.36.0
iniconfig==2.1.0
isodate==0.7.2
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kafka-python==2.0.0
kiwisolver==1.4.9
kubernetes==33.1.0
langchain==0.3.27
langchain-anthropic==0.3.19
langchain-community==0.3.29
langchain-core==0.3.75
langchain-openai==0.3.32
langchain-text-splitters==0.3.11
langcodes==3.5.0
langgraph==0.6.6
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.4
langsmith==0.4.21
language_data==1.3.0
limits==5.5.0
linkify-it-py==2.0.3
locket==1.0.0
mando==0.7.1
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.6
mccabe==0.7.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
merklelib==1.0
minio==7.2.16
mmh3==5.2.0
mpmath==1.3.0
msal==1.33.0
msal-extensions==1.3.1
msgpack==1.1.1
multidict==6.6.4
murmurhash==1.0.13
mypy==1.17.1
mypy_extensions==1.1.0
narwhals==2.2.0
nats-py==2.11.0
neo4j==5.28.2
networkx==3.4.2
nltk==3.9.1
numpy==2.2.6
oauthlib==3.3.1
openai==1.107.1
opencv-python==4.12.0.88
opentelemetry-api==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
ordered-set==4.1.0
orjson==3.11.4
ormsgpack==1.11.0
packaging==25.0
pamqp==3.3.0
pandas==2.3.3
parsimonious==0.10.0
partd==1.4.2
pathspec==0.12.1
pendulum==3.1.0
phonenumbers==9.0.17
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
preshed==3.0.10
presidio_analyzer==2.2.360
presidio_anonymizer==2.2.360
prometheus_client==0.23.1
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.3
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.10.6
pydantic-settings==2.11.0
pydantic_core==2.27.2
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.1.0
pyparsing==3.2.5
pytest==8.4.2
pytest-asyncio==1.1.0
pytest-benchmark==5.1.0
pytest-mock==3.15.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pywin32==311
PyYAML==6.0.3
reactivex==4.0.4
redis==7.0.1
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
requests-file==3.0.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rlp==4.1.0
rpds-py==0.28.0
rsa==4.9.1
s3transfer==0.13.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
smart_open==7.4.2
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.44
srsly==2.5.1
starlette==0.47.3
stevedore==5.5.0
sympy==1.14.0
tabulate==0.9.0
tenacity==8.5.0
thinc==8.3.6
tiktoken==0.12.0
tldextract==5.3.0
toml==0.10.2
tomli==2.3.0
toolz==1.1.0
torch==2.9.0
tqdm==4.67.1
typeguard==4.4.4
typer==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uc-micro-py==1.0.3
uritemplate==4.2.0
url-normalize==2.2.1
urllib3==2.5.0
uvicorn==0.34.0
uvicorn-worker==0.3.0
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.1.1
weasel==0.4.1
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
